More fixes

+++ main.py
@@ -1,3 +1,4 @@
+from genericpath import isfile
 import os
 import uvloop
 from resources import Design
@@ -9,15 +10,20 @@ from aioconsole import aprint
 uvloop.install()
 
 async def main():
-    os.remove("./tokens.txt")
+    if os.path.isfile("./false_tokens.txt"):
+        os.remove("./false_tokens.txt")
+
+    if os.path.isfile("./tokens.txt"):
+        os.remove("./tokens.txt")
+
     repl = ReplIt()
     await Design.ascii()
-    id = await repl.get_id("/@ir0kforever/scalic-selfbot")
+    id = await repl.get_id("/@W1zz3d/Venom-Selfbot-2021-New-Selfbot")
     urls, ids = await repl.get_forks(id)
     await aprint(len(urls))
     full = []
-    for url, id in zip(urls, ids):
 
+    for url, id in zip(urls, ids):
         await repl.get_zip(url, id)
         tokens = await repl.search_zip(id)
         full += tokens
@@ -25,10 +31,25 @@ async def main():
 
     for token in full:
         await aprint(token)
-    with open('./tokens.txt', 'a+') as f:
+    with open('./false_tokens.txt', 'a+') as f:
         for token in full:
             f.write(f"{token}\n")
 
+    lines_seen = set()
+    outfile = open("tokens.txt", "w")
+    for line in open("false_tokens.txt", "r"):
+        if line not in lines_seen:
+            outfile.write(line)
+            lines_seen.add(line)
+    outfile.close()
+    # check()
+
+
+
+def check():
+    repl = ReplIt()
+    tokens = open("tokens.txt").read().splitlines()
+    repl.check_tokens(tokens)
